code refactoring

+++ ms-orders/src/main/java/com/compassuol/sp/challenge/msorders/controller/OrderController.java
@@ -31,7 +31,7 @@ public class OrderController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getOrderById(@PathVariable int id) {
+    public ResponseEntity<?> getOrderById(@PathVariable Long id) {
         if (id <= 0) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("ID de pedido inválido.");
         }
